feat: Implement ES modules
